+++ pkgs/dart_mcp/example/tools_client.dart
@@ -2,7 +2,9 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-// A client that connects to a server and exercises the tools API.
+/// A client that connects to a server and exercises the tools API.
+library;
+
 import 'dart:async';
 import 'dart:io';